resolves #12 Moving dynamic page Drawer to its own component reuitilizable
